+++ .size-limit.js
@@ -234,7 +234,7 @@ module.exports = [
     import: createImport('init'),
     ignore: ['next/router', 'next/constants'],
     gzip: true,
-    limit: '46.5 KB',
+limit: '46.5 KB',
   },
   // SvelteKit SDK (ESM)
   {