+++ dev-packages/e2e-tests/test-applications/browser-webworker-vite/vite.config.ts
@@ -1,10 +1,15 @@
 import { sentryVitePlugin } from '@sentry/vite-plugin';
+import { resolve } from 'path';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
   build: {
     sourcemap: 'hidden',
-    envPrefix: ['PUBLIC_'],
+    rollupOptions: {
+      input: {
+        main: resolve(__dirname, 'index.html'),
+      },
+    },
   },
 
   plugins: [
@@ -26,6 +31,4 @@ export default defineConfig({
       }),
     ],
   },
-
-  envPrefix: ['PUBLIC_'],
 });